#1221 split ComposeMessageService
