+++ FlowCrypt/Functionality/Services/Compose Message Service/ComposeMessageRecipient.swift
@@ -1,0 +1,20 @@
+//
+//  ComposeMessageRecipient.swift
+//  FlowCrypt
+//
+//  Created by Roma Sosnovsky on 17/12/21
+//  Copyright © 2017-present FlowCrypt a. s. All rights reserved.
+//
+
+import Foundation
+
+struct ComposeMessageRecipient {
+    let email: String
+    var state: RecipientState
+}
+
+extension ComposeMessageRecipient: Equatable {
+    static func == (lhs: ComposeMessageRecipient, rhs: ComposeMessageRecipient) -> Bool {
+        return lhs.email == rhs.email
+    }
+}
